Added the "Godkiller Ascended"
She had a long way to the stat she is in now.

Funfact: I started thinking about this concept when I first found the connection between Shuriman and Targonian ascension. I always wanted to write a concept about "an Ascended to end all Ascended", but she ended up as a forcefully abused demigod with a portion of Asols power.